pyproject.toml
[project]
name = "rna-intel"
version = "0.1.0"
description = "RNA State Intelligence Platform - Phase 1"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pandas>=2.1.3",
    "numpy>=1.26.2",
    "torch>=2.1.1",
    "scikit-learn>=1.3.2",
    "scanpy>=1.9.3",
    "anndata>=0.9.2",
    "umap-learn>=0.5.5",
    "plotly>=5.18.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 100
